refactor: streamline NPM release workflow generation by consolidating job creation
