+++ crates/forge_ci/src/workflows/release_homebrew.rs
@@ -1,12 +1,14 @@
 use gh_workflow_tailcall::generate::Generate;
 use gh_workflow_tailcall::*;
 
-use crate::jobs::{self, create_build_release_job_for_publishing, create_homebrew_release_job};
+use crate::jobs::{create_build_release_job_for_publishing, create_homebrew_release_job};
 use crate::matrix;
 
 /// Generate homebrew release workflow
 pub fn generate_homebrew_workflow() {
-    let mut homebrew_workflow = Workflow::default()
+    let build_job = create_build_release_job_for_publishing(matrix::create_matrix());
+    let homebrew_release_job = create_homebrew_release_job().add_needs(build_job.clone());
+    let homebrew_workflow = Workflow::default()
         .name("Homebrew Release")
         .on(Event {
             release: Some(Release { types: vec![ReleaseType::Published] }),
@@ -16,12 +18,7 @@ pub fn generate_homebrew_workflow() {
             Permissions::default()
                 .contents(Level::Write)
                 .pull_requests(Level::Write),
-        );
-
-    let build_job = create_build_release_job_for_publishing(matrix::create_matrix());
-    let homebrew_release_job = create_homebrew_release_job().add_needs(build_job.clone());
-
-    homebrew_workflow = homebrew_workflow
+        )
         .add_job("build-release", build_job)
         .add_job("homebrew_release", homebrew_release_job);
